Split HIR ownership helpers
Move move/projection checks, borrow-region origin tracing, borrowed-slice detection, and active borrow counters out of the HIR monolith into hir/ownership.rs.

Update compiler source monolith reporting and ratchet ceilings for the new compiler.hir ownership module.

Refs #1254.
